Core/Worldserver: Split Worldserver services into separate files (#29555)
