+++ src/server/bnetserver/Services/AccountService.cpp
@@ -16,7 +16,6 @@
  */
 
 #include "AccountService.h"
-#include "Session.h"
 
 Battlenet::Services::Account::Account(Session* session) : AccountService(session)
 {
