+++ src/server/bnetserver/Services/AuthenticationService.cpp
@@ -16,7 +16,6 @@
  */
 
 #include "AuthenticationService.h"
-#include "Session.h"
 
 Battlenet::Services::Authentication::Authentication(Session* session) : AuthenticationService(session)
 {
