+++ src/server/bnetserver/Services/ConnectionService.cpp
@@ -17,7 +17,6 @@
 
 #include "ConnectionService.h"
 #include "Duration.h"
-#include "Session.h"
 #include "Util.h"
 #include "BattlenetRpcErrorCodes.h"
