+++ src/server/bnetserver/Services/ConnectionService.h
@@ -37,7 +37,6 @@ namespace Battlenet
             uint32 HandleConnect(connection::v1::ConnectRequest const* request, connection::v1::ConnectResponse* response, std::function<void(ServiceBase*, uint32, ::google::protobuf::Message const*)>& continuation) override;
             uint32 HandleKeepAlive(NoData const* request) override;
             uint32 HandleRequestDisconnect(connection::v1::DisconnectRequest const* request) override;
-
         };
     }
 }
