+++ src/server/game/Services/WorldserverService.h
@@ -19,22 +19,6 @@
 #define WorldserverService_h__
 
 #include "WorldSession.h"
-#include "account_service.pb.h"
-#include "authentication_service.pb.h"
-#include "challenge_service.pb.h"
-#include "club_listener.pb.h"
-#include "club_membership_listener.pb.h"
-#include "club_membership_service.pb.h"
-#include "club_service.pb.h"
-#include "connection_service.pb.h"
-#include "friends_service.pb.h"
-#include "game_utilities_service.pb.h"
-#include "presence_listener.pb.h"
-#include "presence_service.pb.h"
-#include "report_service.pb.h"
-#include "api/client/v2/report_service.pb.h"
-#include "resource_service.pb.h"
-#include "user_manager_service.pb.h"
 
 namespace bgs { namespace protocol { } }
 using namespace bgs::protocol;
@@ -55,24 +39,6 @@ namespace Battlenet
         std::string GetCallerInfo() const override { return _session->GetPlayerInfo(); }
 
         WorldSession* _session;
-    };
-
-    class GameUtilitiesService : public WorldserverService<game_utilities::v1::GameUtilitiesService>
-    {
-        typedef WorldserverService<game_utilities::v1::GameUtilitiesService> BaseService;
-
-    public:
-        GameUtilitiesService(WorldSession* session);
-
-        uint32 HandleProcessClientRequest(game_utilities::v1::ClientRequest const* request, game_utilities::v1::ClientResponse* response, std::function<void(ServiceBase*, uint32, ::google::protobuf::Message const*)>& continuation) override;
-        uint32 HandleGetAllValuesForAttribute(game_utilities::v1::GetAllValuesForAttributeRequest const* request, game_utilities::v1::GetAllValuesForAttributeResponse* response, std::function<void(ServiceBase*, uint32, ::google::protobuf::Message const*)>& continuation) override;
-
-    private:
-        using ClientRequestHandler = uint32(GameUtilitiesService::*)(std::unordered_map<std::string, Variant const*> const&, game_utilities::v1::ClientResponse*);
-        static std::unordered_map<std::string, ClientRequestHandler> const ClientRequestHandlers;
-
-        uint32 HandleRealmListRequest(std::unordered_map<std::string, Variant const*> const& params, game_utilities::v1::ClientResponse* response);
-        uint32 HandleRealmJoinRequest(std::unordered_map<std::string, Variant const*> const& params, game_utilities::v1::ClientResponse* response);
     };
 }
