+++ src/server/game/Services/WorldserverServiceDispatcher.cpp
@@ -16,7 +16,6 @@
  */
 
 #include "WorldserverServiceDispatcher.h"
-#include "WorldserverService.h"
 
 Battlenet::WorldserverServiceDispatcher::WorldserverServiceDispatcher()
 {
@@ -26,7 +25,7 @@ Battlenet::WorldserverServiceDispatcher::WorldserverServiceDispatcher()
     AddService<WorldserverService<club::v1::ClubService>>();
     AddService<WorldserverService<connection::v1::ConnectionService>>();
     AddService<WorldserverService<friends::v1::FriendsService>>();
-    AddService<GameUtilitiesService>();
+    AddService<Services::GameUtilitiesService>();
     AddService<WorldserverService<presence::v1::PresenceService>>();
     AddService<WorldserverService<report::v1::ReportService>>();
     AddService<WorldserverService<report::v2::ReportService>>();
