+++ src/server/game/Services/WorldserverServiceDispatcher.h
@@ -21,7 +21,6 @@
 #include "WorldSession.h"
 #include "MessageBuffer.h"
 #include "Log.h"
-#include "Common.h"
 #include "account_service.pb.h"
 #include "authentication_service.pb.h"
 #include "challenge_service.pb.h"
@@ -31,7 +30,7 @@
 #include "club_service.pb.h"
 #include "connection_service.pb.h"
 #include "friends_service.pb.h"
-#include "game_utilities_service.pb.h"
+#include "WorldserverGameUtilitiesService.h"
 #include "presence_listener.pb.h"
 #include "presence_service.pb.h"
 #include "report_service.pb.h"
